frontend/requirements.txt
streamlit==1.31.1
requests==2.31.0
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
streamlit-chat==0.1.1
streamlit-option-menu==0.3.12
streamlit-extras==0.3.5
plotly==5.18.0
pandas==2.2.0
numpy==1.26.3 